change types in mqtt.h to correspond to the .c file

+++ sep4_drivers/src/mqtt.h
@@ -1,5 +1,5 @@
 #include <stddef.h>
-int create_mqtt_connect_packet(unsigned char *buf, size_t buflen);
-int create_mqtt_transmit_packet(char *topic, unsigned char *payload,
+size_t create_mqtt_connect_packet(unsigned char *buf, size_t buflen);
+size_t create_mqtt_transmit_packet(char *topic, unsigned char *payload,
                                 unsigned char *buf, size_t buflen);
-int create_mqtt_disconnect_packet(unsigned char *buf, size_t buflen);
+size_t create_mqtt_disconnect_packet(unsigned char *buf, size_t buflen);
